light: convert snnmptrapd helper into a plugin, thus eliminating the need of extra import statements
