+++ tests/python_functional/src/helpers/snmptrapd/SNMPTestParams.py
@@ -20,101 +20,7 @@
 # COPYING for details.
 #
 #############################################################################
-import os
-import re
-
 import pytest
-from pathlib2 import Path
-from psutil import TimeoutExpired
-
-import src.testcase_parameters.testcase_parameters as tc_parameters
-from src.common.blocking import wait_until_true
-from src.driver_io.file.file_io import FileIO
-from src.executors.process_executor import ProcessExecutor
-from src.syslog_ng_config.statements.destinations.destination_reader import DestinationReader
-
-
-class SNMPtrapd(object):
-    TRAP_LOG_PREFIX = 'LIGHT_TEST_SNMP_TRAP_RECEIVED:'
-
-    def __init__(self, port):
-        self.snmptrapd_proc = None
-        self.port = port
-
-        self.snmptrapd_log = Path(tc_parameters.WORKING_DIR, "snmptrapd_log")
-        self.snmptrapd_stdout_path = Path(tc_parameters.WORKING_DIR, "snmptrapd_stdout")
-        self.snmptrapd_stderr_path = Path(tc_parameters.WORKING_DIR, "snmptrapd_stderr")
-
-    def wait_for_snmptrapd_log_creation(self):
-        return self.snmptrapd_log.exists()
-
-    def wait_for_snmptrapd_startup(self):
-        return "NET-SNMP version" in self.snmptrapd_log.read_text()
-
-    def start(self):
-        if self.snmptrapd_proc is not None:
-            return
-
-        self.snmptrapd_proc = ProcessExecutor().start(
-            [
-                "snmptrapd", "-f",
-                "--disableAuthorization=yes",
-                "-C",
-                "-m ALL",
-                "-A",
-                "-Ddump",
-                "-On",
-                "--doNotLogTraps=no",
-                "--authCommunity=log public",
-                self.port,
-                "-d",
-                "-Lf", os.path.relpath(str(self.snmptrapd_log)),
-                "-F", "{}%v\n".format(self.TRAP_LOG_PREFIX),
-            ],
-            self.snmptrapd_stdout_path,
-            self.snmptrapd_stderr_path,
-        )
-        wait_until_true(self.wait_for_snmptrapd_log_creation)
-        wait_until_true(self.wait_for_snmptrapd_startup)
-        return self.snmptrapd_proc.is_running()
-
-    def stop(self):
-        if self.snmptrapd_proc is None:
-            return
-
-        self.snmptrapd_proc.terminate()
-        try:
-            self.snmptrapd_proc.wait(4)
-        except TimeoutExpired:
-            self.snmptrapd_proc.kill()
-
-        self.snmptrapd_proc = None
-
-    def get_port(self):
-        return self.port
-
-    def get_traps(self):
-        file_reader = DestinationReader(FileIO)
-        logs = file_reader.read_all_logs(self.snmptrapd_log)
-        trap_list = []
-        for log_line in logs:
-            res = re.match('({})(.*)'.format(self.TRAP_LOG_PREFIX), log_line)
-            if (res):
-                trap_list += res.group(2).rstrip().split("\t")
-        return sorted(trap_list)
-
-    def get_raw_traps(self):
-        file_reader = DestinationReader(FileIO)
-        return file_reader.read_all_logs(self.snmptrapd_log)
-
-
-@pytest.fixture
-def snmptrapd(port_allocator):
-    server = SNMPtrapd(port_allocator())
-    server.start()
-    yield server
-    server.stop()
-
 
 class SNMPTestParams(object):
     def __init__(self):
